change all double to float to be consistent with CUDA side

+++ include/mid_plan/DijkstraMap.h
@@ -9,16 +9,16 @@ struct nodeInfo
 {
     bool inClosed;
     bool inQ;
-    double g;
-    double h;
+    float g;
+    float h;
     CUDA_GEO::coord c;
     nodeInfo* ptr2parent;
-    std::multiset<std::pair<double, nodeInfo*>>::iterator it;
+    std::multiset<std::pair<float, nodeInfo*>>::iterator it;
     nodeInfo():
         inClosed(false),
         inQ(false),
-        g(std::numeric_limits<double>::infinity()),
-        h(std::numeric_limits<double>::infinity()),
+        g(std::numeric_limits<float>::infinity()),
+        h(std::numeric_limits<float>::infinity()),
         ptr2parent(nullptr)
     {
 
@@ -29,7 +29,7 @@ class DijkstraMap : public MapBase
 {
 public:
     DijkstraMap(int maxX, int maxY, int maxZ);
-    double obsCostAt(CUDA_GEO::coord s, float default_value, bool &occupied,
+    float obsCostAt(CUDA_GEO::coord s, float default_value, bool &occupied,
                      const CUDA_GEO::coord *crd_shift = nullptr, SeenDist *last_val_map = nullptr, float obstacle_dist = 1.0) const;
     bool isSeen(const CUDA_GEO::coord & s, const bool default_value) const;
 
@@ -39,15 +39,15 @@ public:
     int getMaxX() {return _w;}
     int getMaxY() {return _h;}
     int getMaxZ() {return _d;}
-    double getCost2Come(const CUDA_GEO::coord & s, const double &default_value) const;
+    float getCost2Come(const CUDA_GEO::coord & s, const float &default_value) const;
 
     void copyIdData(const cpc_aux_mapping::grid_map::ConstPtr &msg);
     void copyEdtData(const cpc_aux_mapping::grid_map::ConstPtr &msg);
 
-    std::vector<CUDA_GEO::coord> AStar2D(const CUDA_GEO::coord &goal, const CUDA_GEO::coord &start, bool reached_free_zone, double &length,
+    std::vector<CUDA_GEO::coord> AStar2D(const CUDA_GEO::coord &goal, const CUDA_GEO::coord &start, bool reached_free_zone, float &length,
                                        const CUDA_GEO::coord *crd_shift = nullptr, SeenDist *last_val_map = nullptr);
     CUDA_GEO::coord findTargetCoord(const std::vector<CUDA_GEO::coord> &path);
-    int calcTgtHeightCoord(double tgt_height);
+    int calcTgtHeightCoord(float tgt_height);
     std::vector<CUDA_GEO::coord> rayCast(const CUDA_GEO::coord &p0Index, const CUDA_GEO::coord &p1Index, float limit_radius = -1);
     std::vector<CUDA_GEO::pos> findSplitCoords(const std::vector<CUDA_GEO::coord> &path);
     bool checkTopo(const std::vector<CUDA_GEO::coord> &path_a,const std::vector<CUDA_GEO::coord> &path_b);
@@ -69,22 +69,22 @@ private:
         return s.z*_w*_h+s.y*_w+s.x;
     }
 
-    inline double dist(const CUDA_GEO::coord & c1, const CUDA_GEO::coord & c2)
+    inline float dist(const CUDA_GEO::coord & c1, const CUDA_GEO::coord & c2)
     {
         CUDA_GEO::coord c = c1-c2;
-        return sqrt(static_cast<double>(c.square()));
+        return sqrt(static_cast<float>(c.square()));
     }
 
-    inline double point2lineDist(const CUDA_GEO::coord & c1, const CUDA_GEO::coord & c2, const CUDA_GEO::coord & c0)
+    inline float point2lineDist(const CUDA_GEO::coord & c1, const CUDA_GEO::coord & c2, const CUDA_GEO::coord & c0)
     {
         CUDA_GEO::coord a = c1-c0;
         CUDA_GEO::coord b = c2-c1;
         int a_dot_b = a.x*b.x + a.y*b.y + a.z*b.z;
 
         if (b.square() == 0)
-            return sqrt(static_cast<double>(a.square()));
+            return sqrt(static_cast<float>(a.square()));
 
-        return sqrt(static_cast<double>(a.square()*b.square() - a_dot_b*a_dot_b)/static_cast<double>(b.square()));
+        return sqrt(static_cast<float>(a.square()*b.square() - a_dot_b*a_dot_b)/static_cast<float>(b.square()));
     }
 
 public:
